move structs/classes to their own files
implement lazy evaluation of font measuring by making their data optional and private with getters which do the evaluation on demand
