+++ Source/UnicodeBrowser/UnicodeBrowserWidget.h
@@ -3,13 +3,14 @@
 #pragma once
 
 #include "CoreMinimal.h"
+#include "UnicodeBrowserOptions.h"
+#include "UnicodeBrowserRow.h"
 
 #include "Fonts/UnicodeBlockRange.h"
 
 #include "Widgets/SCompoundWidget.h"
 #include "Widgets/Views/SListView.h"
 
-#include "UnicodeBrowserWidget.generated.h"
 
 class UFont;
 class SScrollBox;
@@ -36,65 +37,6 @@ namespace UnicodeBrowser
 	static TArray<EUnicodeBlockRange> GetSymbolRanges();
 }
 
-class FUnicodeBrowserRow : public TSharedFromThis<FUnicodeBrowserRow>
-{
-public:
-	FUnicodeBrowserRow() = default;
-
-	FUnicodeBrowserRow(int32 CodePointIn, TOptional<EUnicodeBlockRange> BlockRangeIn) : Codepoint(CodePointIn), BlockRange(BlockRangeIn) {}
-
-	int32 Codepoint = 0;
-	FString Character = "";
-	TOptional<EUnicodeBlockRange> BlockRange;
-	FFontData FontData;
-	float ScalingFactor = 1.0f;
-	FVector2D Measurements;
-	bool bCanLoadCodepoint = false;
-
-	friend bool operator==(FUnicodeBrowserRow const& Lhs, FUnicodeBrowserRow const& RHS)
-	{
-		return Lhs.Codepoint == RHS.Codepoint
-			&& Lhs.Character == RHS.Character
-			&& Lhs.BlockRange == RHS.BlockRange
-			&& Lhs.FontData == RHS.FontData
-			&& Lhs.ScalingFactor == RHS.ScalingFactor
-			&& Lhs.Measurements == RHS.Measurements
-			&& Lhs.bCanLoadCodepoint == RHS.bCanLoadCodepoint;
-	}
-
-	friend bool operator!=(FUnicodeBrowserRow const& Lhs, FUnicodeBrowserRow const& RHS) { return !(Lhs == RHS); }
-};
-
-UCLASS(Hidden, BlueprintType, EditInlineNew, DefaultToInstanced, DisplayName = "Font Options")
-class UNICODEBROWSER_API UUnicodeBrowserOptions : public UObject
-{
-	GENERATED_BODY()
-
-public:
-	static TSharedRef<class IDetailsView> MakePropertyEditor(UUnicodeBrowserOptions* Options);
-	UPROPERTY(EditAnywhere, meta=(ShowOnlyInnerProperties), Transient)
-	FSlateFontInfo Font = FCoreStyle::GetDefaultFontStyle("Regular", 18);
-
-	UPROPERTY(EditAnywhere)
-	int32 NumCols = 24;
-
-	UPROPERTY(EditAnywhere)
-	bool bShowMissing = false;
-
-	UPROPERTY(EditAnywhere)
-	bool bShowZeroSize = false;
-
-	virtual void PostInitProperties() override;
-
-	virtual void PostEditChangeProperty(struct FPropertyChangedEvent& PropertyChangedEvent) override;
-
-	DECLARE_MULTICAST_DELEGATE(FOnNumColsChangedDelegate);
-	FOnNumColsChangedDelegate OnChanged;
-};
-
-/**
- * 
- */
 class SUnicodeBrowserWidget : public SCompoundWidget
 {
 public:
@@ -105,7 +47,6 @@ public:
 	
 protected:
 	TArrayView<FUnicodeBlockRange const> Ranges; // all known unicode ranges
-	
 	TMap<EUnicodeBlockRange, TArray<TSharedPtr<FUnicodeBrowserRow>>> Rows;	
 	TMap<EUnicodeBlockRange, TSharedPtr<SExpandableArea>> RangeWidgets;
 	TMap<EUnicodeBlockRange, TSharedPtr<SUniformGridPanel>> RangeWidgetsGrid;
@@ -132,21 +73,6 @@ protected:
 	void RebuildGridColumns(::FUnicodeBlockRange Range, TSharedRef<SUniformGridPanel> const GridPanel) const;
 
 	
-	void UpdateFromFont();
+	void UpdateFromFont(struct FPropertyChangedEvent* PropertyChangedEvent = nullptr);
 };
 
-
-class SUnicodeCharacterInfo: public SCompoundWidget
-{
-public:
-	SLATE_BEGIN_ARGS(SUnicodeCharacterInfo) {}
-		SLATE_ATTRIBUTE(TSharedPtr<FUnicodeBrowserRow>, Row);
-	SLATE_END_ARGS()
-
-	void Construct(FArguments const& InArgs);
-	
-	TAttribute<TSharedPtr<FUnicodeBrowserRow>> Row;
-	TSharedPtr<FUnicodeBrowserRow> GetRow() const;
-	
-	void SetRow(TSharedPtr<FUnicodeBrowserRow> InRow);
-};
